WIP: only grab cls from embeddings, move haystack modules to haystack_integrations
